connected commiting and logging
