+++ app/actions/history.js
@@ -1,0 +1,1 @@
+export const UPDATE_HISTORY_STATE = "UPDATE_HISTORY_STATE";
