+++ app/components/ProseEditor/Editor.js
@@ -1,8 +1,6 @@
 import React from "react";
 import { connect } from "react-redux";
 import { Editor, RichUtils, getDefaultKeyBinding } from "draft-js";
-import { Button } from "reactstrap";
-import { convertToRaw } from "draft-js";
 
 import InlineStyleControls from "./InlineStyleControls";
 import BlockStyleControls from "./BlockStyleControls";
@@ -68,7 +66,7 @@ export class ProseEditor extends React.Component {
   }
 
   render() {
-    const { editorState, currentFile, gitAbstractions } = this.props;
+    const { editorState } = this.props;
 
     // If the user changes block type before entering any text, we can
     // either style the placeholder or hide it. Let's just hide it now.
@@ -86,16 +84,6 @@ export class ProseEditor extends React.Component {
     }
     return (
       <div className={styles.root}>
-        <Button
-          onClick={() =>
-            gitAbstractions.saveFile(
-              currentFile,
-              convertToRaw(editorState.getCurrentContent())
-            )
-          }
-        >
-          Save
-        </Button>
         <BlockStyleControls
           editorState={editorState}
           onToggle={this.toggleBlockType}
@@ -146,11 +134,7 @@ function getBlockStyle(block) {
   }
 }
 
-const mapStateToProps = ({
-  editor: { editorState },
-  currentFile,
-  gitAbstractions
-}) => ({ editorState, currentFile, gitAbstractions });
+const mapStateToProps = ({ editor: { editorState } }) => ({ editorState });
 
 const mapDispatchToProps = dispatch => ({
   onSaveEditorState: editorState => {
