+++ app/containers/EditorPage.js
@@ -1,15 +1,24 @@
 import React from "react";
 import { connect } from "react-redux";
+import { convertToRaw } from "draft-js";
+
 import EditorPanel from "app/components/EditorPanel";
 import TitleBar from "app/components/TitleBar";
 import FileList from "app/components/FileList";
 import History from "app/components/History";
 import CheckpointForm from "app/components/Forms/CheckpointForm";
+import { Button } from "reactstrap";
 
 import styles from "./EditorPage.scss";
 
 const EditorPage = props => {
-  const { showHistory, showFileList } = props;
+  const {
+    editorState,
+    showHistory,
+    showFileList,
+    gitAbstractions,
+    currentFile
+  } = props;
   return (
     <div>
       <TitleBar />
@@ -24,11 +33,25 @@ const EditorPage = props => {
             {/* TODO: should this onSubmit be in a different place? */}
             <CheckpointForm
               onSubmit={({ commitMessage }) =>
-                console.log("Submitted new checkpoint: ", commitMessage)
+                gitAbstractions.saveFile(
+                  currentFile,
+                  convertToRaw(editorState.getCurrentContent()),
+                  commitMessage
+                )
               }
             />
+            <Button
+              onClick={() =>
+                gitAbstractions.saveFile(
+                  currentFile,
+                  convertToRaw(editorState.getCurrentContent())
+                )
+              }
+            >
+              Save
+            </Button>
             {/* TODO: remove prop */}
-            <EditorPanel isEditable={true} />
+            <EditorPanel editorState={editorState} isEditable={true} />
           </div>
           {showHistory && (
             <div className={`${styles.right} col-2`}>
@@ -41,12 +64,18 @@ const EditorPage = props => {
   );
 };
 
-const mapStateToProps = state => {
-  return {
-    showFileList: state.view.showFileList,
-    showHistory: state.view.showHistory
-  };
-};
+const mapStateToProps = ({
+  view: { showHistory, showFileList },
+  editor: { editorState },
+  gitAbstractions,
+  currentFile
+}) => ({
+  showFileList,
+  showHistory,
+  gitAbstractions,
+  currentFile,
+  editorState
+});
 
 const WithEditorPage = connect(mapStateToProps)(EditorPage);
