+++ app/git-abs/git-abs.js
@@ -124,8 +124,14 @@ class GitAbs {
    */
   getVersions = async fileName => {
     // use metedata object to get version names for the given fileName
-    const obj = await this.metadata.getAllVersions(fileName);
-    return Object.keys(obj);
+    // const obj = await this.metadata.getAllVersions(fileName);
+    // return Object.keys(obj);
+    const now = new Date().toLocaleString("en-US", { timeZone: "UTC" });
+    return [
+      { message: "first commit", date: now },
+      { message: "second commit", date: now },
+      { message: "day 68. still doesn't work. i'm giving up", date: now }
+    ];
   };
 
   /**
